refactor: move fetch logic into Home component
