+++ src/routes/home-page/Home.tsx
@@ -1,20 +1,45 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import PhotoCard from '../../components/product-card/PhotoCard';
-import { SearchBar, SearchBarProps } from '../../components/search-bar/SearchBar';
-import { Photo } from '../../interfaces/response';
+import { SearchBar } from '../../components/search-bar/SearchBar';
+import { Photo, SearchResult } from '../../interfaces/response';
 import styles from './home-page.module.scss';
 import Modal from '../../components/Modal';
+import UnsplashLoader from '../../libs/loader';
 
-interface HomePageProps extends SearchBarProps {
-  photos: Photo[];
-  isDataLoaded: boolean;
-}
-
-const HomePage = (props: HomePageProps) => {
+const HomePage = () => {
   const [isOpen, setIsOpen] = useState(false);
   const [photoToShow, setPhotoToShow] = useState<Photo>();
 
-  const cards = props.photos.map((item) => (
+  const storageKey = 'inputValue';
+  const storedValue = localStorage.getItem(storageKey) || '';
+  const [photos, setPhotos] = useState<Photo[]>([]);
+  const [isDataLoaded, setIsDataLoaded] = useState(false);
+  const [search, setSearch] = useState(storedValue);
+  const [prevSearch, setPrevSearch] = useState(storedValue);
+
+  const updateSearchState = (searchStr: string) => {
+    setPrevSearch(search);
+    setSearch(searchStr);
+    localStorage.setItem(storageKey, searchStr);
+    setIsDataLoaded(false);
+  };
+
+  useEffect(() => {
+    if (photos.length === 0 || search !== prevSearch) {
+      UnsplashLoader.getPhotos(search)
+        .then((resp) => resp.json())
+        .then((data: Photo[] | SearchResult) => {
+          setIsDataLoaded(true);
+          if ('results' in data) {
+            setPhotos(data.results);
+          } else {
+            setPhotos(data);
+          }
+        });
+    }
+  }, [photos.length, prevSearch, search]);
+
+  const cards = photos.map((item) => (
     <PhotoCard
       key={item.id.toString()}
       img={item.urls.thumb}
@@ -28,7 +53,7 @@ const HomePage = (props: HomePageProps) => {
 
   return (
     <>
-      <SearchBar updSearch={props.updSearch} search={props.search} />
+      <SearchBar updSearch={updateSearchState} search={search} />
       <Modal
         handleClose={() => {
           setIsOpen(false);
@@ -37,7 +62,7 @@ const HomePage = (props: HomePageProps) => {
         photo={photoToShow as Photo}
       />
       <section className={styles.section}>
-        {props.isDataLoaded ? <div className={styles.cards}>{cards}</div> : <h2>Loading...</h2>}
+        {isDataLoaded ? <div className={styles.cards}>{cards}</div> : <h2>Loading...</h2>}
       </section>
     </>
   );
